fix the bug: the build dir isnot exsited.

+++ Linux/mail/build.sh
@@ -1,9 +1,16 @@
 #!/bin/bash 
 
-rm -rf ./build/*
+if [ ! -d "./build" ];then 
+	mkdir build 
+else 
+	rm -rf ./build/*
+fi  
 
 cd ./build 
+
 cmake ..
+
 echo "start make...."
+
 make
